+++ src/test/java/com/ejoongseok/wmslive/location/feature/AssignLocationLPNTest.java
@@ -1,13 +1,8 @@
 package com.ejoongseok.wmslive.location.feature;
 
-import com.ejoongseok.wmslive.inbound.domain.LPN;
-import com.ejoongseok.wmslive.inbound.domain.LPNRepository;
-import com.ejoongseok.wmslive.location.domain.Location;
-import com.ejoongseok.wmslive.location.domain.LocationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.util.Assert;
 
 class AssignLocationLPNTest {
 
@@ -34,25 +29,4 @@ class AssignLocationLPNTest {
         //TODO locaiton에 locationLPN 목록이 비어있지 않은지 확인한다.
     }
 
-    private class AssignLocationLPN {
-        private LocationRepository locationRepository;
-        private LPNRepository lpnRepository;
-
-        public void request(final Request request) {
-            final Location location = locationRepository.getByLocationBarcode(request.locationBarcode);
-            final LPN lpn = lpnRepository.getByLPNBarcode(request.lpnBarcode);
-
-            location.assignLPN(lpn);
-        }
-
-        public record Request(
-                String locationBarcode,
-                String lpnBarcode) {
-
-            public Request {
-                Assert.hasText(locationBarcode, "로케이션 바코드는 필수입니다.");
-                Assert.hasText(lpnBarcode, "LPN 바코드는 필수입니다.");
-            }
-        }
-    }
 }
